rename model namespace to models
