--- WalletWasabi.Fluent/Controls/TileSize.cs
@@ -1,9 +1,0 @@
-namespace WalletWasabi.Fluent.Controls
-{
-	public enum TileSize
-	{
-		Small,
-		Normal,
-		Wide
-	}
-}
