+++ WalletWasabi.Fluent/Helpers/AuthorizationHelpers.cs
@@ -1,4 +1,4 @@
-using WalletWasabi.Fluent.Model;
+using WalletWasabi.Fluent.Models;
 using WalletWasabi.Fluent.ViewModels.Dialogs.Authorization;
 using WalletWasabi.Wallets;
