+++ WalletWasabi.Fluent/Models/BalanceType.cs
@@ -1,4 +1,4 @@
-namespace WalletWasabi.Fluent.Model
+namespace WalletWasabi.Fluent.Models
 {
 	public enum BalanceType
 	{
