+++ WalletWasabi.Fluent/Models/TileSize.cs
@@ -1,0 +1,9 @@
+namespace WalletWasabi.Fluent.Models
+{
+	public enum TileSize
+	{
+		Medium,
+		Large,
+		Wide
+	}
+}
