+++ WalletWasabi.Fluent/ViewModels/Wallets/Home/Tiles/TilePresetViewModel.cs
@@ -1,4 +1,5 @@
 using WalletWasabi.Fluent.Controls;
+using WalletWasabi.Fluent.Models;
 
 namespace WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles
 {
