+++ WalletWasabi.Fluent/ViewModels/Wallets/WalletViewModel.cs
@@ -11,6 +11,7 @@ using NBitcoin;
 using WalletWasabi.Fluent.ViewModels.Navigation;
 using System.Windows.Input;
 using WalletWasabi.Fluent.Controls;
+using WalletWasabi.Fluent.Models;
 using WalletWasabi.Fluent.ViewModels.Dialogs.Authorization;
 using WalletWasabi.Fluent.ViewModels.Dialogs.Base;
 using WalletWasabi.Fluent.ViewModels.Wallets.Advanced;
@@ -78,9 +79,9 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets
 			{
 				TilePresets = new ObservableCollection<TilePresetViewModel>()
 				{
-					new(0, 0, 1, 1, TileSize.Small),
-					new(0, 0, 1, 1, TileSize.Small),
-					new(0, 0, 1, 1, TileSize.Small)
+					new(0, 0, 1, 1, TileSize.Medium),
+					new(0, 0, 1, 1, TileSize.Medium),
+					new(0, 0, 1, 1, TileSize.Medium)
 				},
 				TilePresetIndex = LayoutIndex
 			};
@@ -88,9 +89,9 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets
 			{
 				TilePresets = new ObservableCollection<TilePresetViewModel>()
 				{
-					new(1, 0, 1, 1, TileSize.Small),
-					new(1, 0, 1, 1, TileSize.Small),
-					new(1, 0, 1, 1, TileSize.Small)
+					new(1, 0, 1, 1, TileSize.Medium),
+					new(1, 0, 1, 1, TileSize.Medium),
+					new(1, 0, 1, 1, TileSize.Medium)
 				},
 				TilePresetIndex = LayoutIndex
 			};
@@ -98,9 +99,9 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets
 			{
 				TilePresets = new ObservableCollection<TilePresetViewModel>()
 				{
-					new(2, 0, 1, 1, TileSize.Small),
-					new(2, 0, 1, 1, TileSize.Small),
-					new(0, 1, 1, 1, TileSize.Normal)
+					new(2, 0, 1, 1, TileSize.Medium),
+					new(2, 0, 1, 1, TileSize.Medium),
+					new(0, 1, 1, 1, TileSize.Large)
 				},
 				TilePresetIndex = LayoutIndex
 			};
@@ -108,9 +109,9 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets
 			{
 				TilePresets = new ObservableCollection<TilePresetViewModel>()
 				{
-					new(3, 0, 1, 1, TileSize.Small),
-					new(0, 1, 1, 1, TileSize.Normal),
-					new(1, 1, 1, 1, TileSize.Normal)
+					new(3, 0, 1, 1, TileSize.Medium),
+					new(0, 1, 1, 1, TileSize.Large),
+					new(1, 1, 1, 1, TileSize.Large)
 				},
 				TilePresetIndex = LayoutIndex
 			};
@@ -118,7 +119,7 @@ namespace WalletWasabi.Fluent.ViewModels.Wallets
 			{
 				TilePresets = new ObservableCollection<TilePresetViewModel>()
 				{
-					new(4, 0, 1, 1, TileSize.Small),
+					new(4, 0, 1, 1, TileSize.Medium),
 					new(1, 1, 2, 1, TileSize.Wide),
 					new(0, 2, 2, 1, TileSize.Wide)
 				},
